nw/sign-language-api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
numpy==1.24.3
opencv-python==4.8.1.78
tensorflow==2.13.0
pillow==10.0.0
python-multipart==0.0.6
aiofiles==23.2.1
pydantic==2.4.2
pyttsx3==2.90
